+++ lib/components/mint-card/index.ts
@@ -1,1 +1,17 @@
 export { MintCard, type MintCardProps } from "./MintCard";
+export {
+  MintErrorStatusFrame,
+  type MintErrorStatusFrameProps,
+} from "./MintErrorStatusFrame";
+export {
+  MintCompletedStatusFrame,
+  type MintCompletedStatusFrameProps,
+} from "./MintCompletedStatusFrame";
+export {
+  IsMintingStatusFrame,
+  type IsMintingStatusFrameProps,
+} from "./IsMintingStatusFrame";
+export {
+  MintEnabledStatusFrame,
+  type MintEnabledStatusFrameProps as MintingEnabledStatusFrameProps,
+} from "./MintingEnabledStatusFrame";
